+++ CryptoCop/Cryptocop.Software.API/Cryptocop.Software.API.Models/Entities/ShoppingCartItem.cs
@@ -8,8 +8,8 @@ namespace Cryptocop.Software.API.Models.Entities
         public int Id { get; set; }
         public int ShoppingCartId { get; set; }
         public string ProductIdentifier { get; set; }
-        public double Quantity { get; set; }
-        public double UnitPrice { get; set; }
+        public float Quantity { get; set; }
+        public float UnitPrice { get; set; }
 
         // Navigation properties
         [ForeignKey("ShoppingCartId")]
